+++ Plugins/FreeRDP/Client/Resource/Translations/FreeRDP_zh_TW.ts
@@ -4,115 +4,126 @@
 <context>
     <name>CConnectFreeRDP</name>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="132"/>
-        <source>freerdp_connect connect to </source>
-        <translation>連接到 </translation>
-    </message>
-    <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="136"/>
-        <source> fail</source>
-        <translation> 失敗</translation>
-    </message>
-    <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="142"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="144"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="164"/>
         <source>Connect to </source>
         <translation>連接到 </translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="555"/>
-        <source>Logon error info [</source>
-        <translation>登錄錯誤信息 [</translation>
+        <location filename="../../ConnectFreeRDP.cpp" line="148"/>
+        <source> fail:</source>
+        <translation>失敗：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="750"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="160"/>
+        <source>Error</source>
+        <translation>錯誤</translation>
+    </message>
+    <message>
+        <location filename="../../ConnectFreeRDP.cpp" line="133"/>
+        <source>Logon to </source>
+        <translation>登錄到 </translation>
+    </message>
+    <message>
+        <location filename="../../ConnectFreeRDP.cpp" line="137"/>
+        <source> fail. Please check that the username and password are correct.</source>
+        <translation>失敗。請檢查用戶名和密碼是否正確。</translation>
+    </message>
+    <message>
+        <location filename="../../ConnectFreeRDP.cpp" line="581"/>
+        <source>FreeRDP logon info: [</source>
+        <translation>FreeRDP 登錄信息：</translation>
+    </message>
+    <message>
+        <location filename="../../ConnectFreeRDP.cpp" line="778"/>
         <source>Verify certificate</source>
         <translation>驗證證書</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="752"/>
-        <location filename="../../ConnectFreeRDP.cpp" line="801"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="780"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="829"/>
         <source>Host: %1; Port: %2</source>
         <translation>主機：%1; 端口：%2</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="753"/>
-        <location filename="../../ConnectFreeRDP.cpp" line="802"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="781"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="830"/>
         <source>Common name: </source>
         <translation>公共名稱：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="754"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="782"/>
         <source>Subject: </source>
         <translation>主題：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="755"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="783"/>
         <source>Issuer: </source>
         <translation>發行者：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="756"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="784"/>
         <source>Fingerprint: </source>
         <translation>指紋：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="762"/>
-        <location filename="../../ConnectFreeRDP.cpp" line="814"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="790"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="842"/>
         <source>Don&apos;t show again</source>
         <translation>不再顯示</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="799"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="827"/>
         <source>Verify changed certificate</source>
         <translation>驗證證書改變</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="803"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="831"/>
         <source>New subject: </source>
         <translation>新的主題：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="804"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="832"/>
         <source>New issuer: </source>
         <translation>新的發行者：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="805"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="833"/>
         <source>New fingerprint: </source>
         <translation>新的指紋：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="806"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="834"/>
         <source>Old subject: </source>
         <translation>舊的主題：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="807"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="835"/>
         <source>Old issuer: </source>
         <translation>舊的發行者：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="808"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="836"/>
         <source>Old fingerprint: </source>
         <translation>舊的指紋：</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="845"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="873"/>
         <source>Consent message</source>
         <translation>同意消息</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="845"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="873"/>
         <source>Service message</source>
         <translation>服務留言</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="858"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="886"/>
         <source>I understand and agree to the terms of this policy (Y/N)</source>
         <translation>我理解並同意本政策的條款（是/否）</translation>
     </message>
     <message>
-        <location filename="../../ConnectFreeRDP.cpp" line="864"/>
+        <location filename="../../ConnectFreeRDP.cpp" line="892"/>
         <source>Gateway message</source>
         <translation>網關消息</translation>
     </message>